Validate Web Profile (#14)
Closes issue #1.

The URL found in the client certificate under "Subject Alternate Name" is used to fetch the user's profile document and the certificate is validated against the certificate data found in the profile. See WebID-TLS (https://www.w3.org/2005/Incubator/webid/spec/tls/).

Added bonus: we're able to display the user's name in the UI having fetched that data from the profile as well.